banksampah

+++ app/Http/Controllers/Admin/UserController.php
@@ -164,9 +164,12 @@ class UserController extends Controller
     public function destroy($id)
     {
         $user = User::find($id);
-        $user->delete();
-
-        alert::success('message', 'User Removed');
+        if ($user->role(['admin', 'pengurus1'])) {
+            alert::error('message', 'Admin User Cannot Be Deleted');
+        } else {
+            $user->delete();
+            alert::success('message', 'User Removed');
+        }
         return redirect('admin/user');
     }
 }
